pyproject.toml
[tool.poetry]
name = "flowagent"
version = "0.1.0"
description = "FlowAgent demo"
authors = ["easonsshi <easonsshi@tencent.com>"]
readme = "README.md"
packages = [
    { include = "fa_core", from = "src" },
    { include = "fa_demo", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.51.2"
Jinja2 = "^3.1.4"
colorama = "^0.4.6"
pyyaml = "^6.0.2"
seaborn = "^0.13.2"
tabulate = "^0.9.0"
wandb = "^0.18.5"
loguru = "^0.7.3"
pydantic = "^2.10.3"
requests = "^2.32.3"
matplotlib = "^3.10.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
tencentcloud-sdk-python = "^3.0.1251"
jwcrypto = "^1.5.6"

[tool.poetry.group.backend.dependencies]
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
python-engineio = "^4.10.1"
python-socketio = "^5.11.4"
pymongo = "^4.10.1"
ipython = "^8.31.0"                     # for tool_python_executor
sympy = "^1.13.3"                       # for tool_calculator (math)
langchain-community = "^0.3.9"          # GoogleSerperAPIWrapper
# langchain-experimental = "^0.3.4"     # use IPython's InteractiveShell instead of langchain-experimental's

[tool.poetry.group.frontend.dependencies]
streamlit = "^1.41.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
